+++ client/src/components/Readme/readme.css
@@ -1,65 +1,30 @@
-/* General Styles */
-body {
-	font-family: Arial, Helvetica, sans-serif;
-	margin: 0;
-	padding: 0;
-	background-color: #f4f4f4;
-}
-
-.container {
+.readme-container {
 	max-width: 800px;
 	margin: 20px auto;
 	padding: 20px;
 	background-color: white;
 	box-shadow: 0 0 10px rgba(0, 0, 0, 0.1);
 	border-radius: 8px;
+	font-family: Arial, Helvetica, sans-serif;
+	background-color: #f4f4f4;
 }
 
-/* Header Styles */
-header {
-	text-align: center;
-	margin-bottom: 20px;
-}
-
-header h1 {
-	font-size: 2.5em;
-	margin-bottom: 10px;
-	color: #333;
-}
-
-header p {
-	font-size: 1.2em;
-	color: #555;
-}
-
-header a {
-	color: #007bff;
-	text-decoration: none;
-	transition: color 0.3s ease;
-}
-
-header a:hover {
-	color: #0056b3;
-}
-
-/* Main Content Styles */
-h2 {
+.readme-container h2 {
 	font-size: 2em;
 	margin-bottom: 10px;
 	color: #333;
 }
 
-p {
+.readme-container p {
 	font-size: 1.1em;
 	line-height: 1.6;
 	color: #555;
 	margin-bottom: 20px;
 }
 
-/* Navigation Links Styles */
 .navig {
 	display: flex;
-	gap: 10px;
+	gap: 12px;
 	margin-bottom: 20px;
 }
